Dev2

+++ database/migrations/2023_08_09_000004_create_number_of_employees_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->bigIncrements('id');
             $table->string('label');
             $table->smallInteger('min');
-            $table->smallInteger('max');
+            $table->bigInteger('max');
 
             $table->timestamps();
         });
